data-acquisition-framework/pyproject.toml
[project]
name = "data-acquisition-framework"
version = "0.5.0"
description = ""
authors = [
    {name = "KN Neuron"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame (>=2.6.1,<3.0.0)",
    "colour (>=0.1.5,<0.2.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[tool.poetry]
packages = [{ from = "src", include = "data_acquisition" }]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pyright = "^1.1.403"
sphinx = "^8.2.3"
sphinx-autobuild = "^2024.10.3"
sphinx-rtd-theme = "^3.0.2"
sphinxcontrib-napoleon = "^0.7"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
